+++ xcode/MAME4iOS/DeviceScreenResolver.m
@@ -13,9 +13,16 @@
 +(DeviceScreenType) resolve {
     UIUserInterfaceIdiom idiom = UIDevice.currentDevice.userInterfaceIdiom;
     CGSize screenSize = UIScreen.mainScreen.bounds.size;
+    #pragma clang diagnostic push
+    #pragma clang diagnostic ignored "-Wdeprecated"
     CGSize windowSize = UIApplication.sharedApplication.keyWindow.bounds.size;
+    #pragma clang diagnostic pop
     CGFloat maxLength = MAX(screenSize.width, screenSize.height);
 
+    assert(idiom != 5);     // 5 is UIUserInterfaceIdiomMac
+    assert(idiom != UIUserInterfaceIdiomUnspecified);
+    assert(idiom == UIUserInterfaceIdiomPad || idiom == UIUserInterfaceIdiomPhone);
+    
     if ( idiom == UIUserInterfaceIdiomPad && !CGSizeEqualToSize(windowSize, CGSizeZero) && !CGSizeEqualToSize(screenSize, windowSize)) {
         // we are on an iPad in SlideOver or SplitScreen mode. pretend to be a generic iPhone or iPad based on aspect.
 
@@ -57,6 +64,7 @@
             return IPAD_GENERIC;
         }
     } else {
+        assert(FALSE);
         return IPAD_GENERIC;
     }
 }