+++ iOS-res/Applications/MAME4iOS.app/whatsnew.md
@@ -7,6 +7,11 @@
 * PINCH and ZOOM to resize the HUD to exacly the size you want.
 * set MAME pause_brightness to 1.0
 * Support for Xbox `GUIDE`, DualShock `PS`, and Steam `STEAM` buttons on iOS 14.
+* Support WideColor (aka DisplayP3) devices.
+* Support SlideOver and SplitScreen on iPad.
+* Support macOS via Catalyst, requires macOS 10.15 (aka Catalina) or higher.
+* Mac menu bar and edge to edge game window.
+* A,B,Y,X,L,R on a hardware keyboard will function as buttons.
 
 # Version 2020.11
 * Added new Vector Shader category for Vector games on both iOS and tvOS.